fix: 통계 반환 데이터 수정

+++ back/dream/src/main/java/com/dream/dream/statistic/dto/StatisticDto.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class StatisticDto {
     private Long memberId;
     private String keyword;
-    private int count;
+    private Long count;
 }
